Fixed nightly code coverage [ci skip]
The code coverage GA was referencing 1.21.x which seem to prevent
the install of gocovmerge.
Also added an early check to see if we can run gocovmerge before
starting the test suite.

+++ scripts/cov.sh
@@ -17,6 +17,12 @@ check_file () {
 export GO111MODULE="on"
 
 go install github.com/wadey/gocovmerge@latest
+# Fail fast by checking if we can run gocovmerge
+gocovmerge
+if [[ $? != 0 ]]; then
+    echo "Unable to run gocovmerge"
+    exit 1
+fi
 
 rm -rf ./cov
 mkdir cov
